Fix build warnings on Raspberry Pi
Resolves compilation errors discovered during Pi deployment.

What:
- Add sys/types.h include to tap.h for ssize_t type
- Cast signedness comparison in fragment.c

Why:
- tap.h: ssize_t type undefined on arm64 gcc without sys/types.h
- fragment.c: Comparing size_t with int triggers -Werror=sign-compare

How:
- tap.h: Add #include <sys/types.h> after stddef.h
- fragment.c: Cast FRAG_HEADER_SIZE + payload_len to (size_t)

Testing:
- Verified clean build on Raspberry Pi 4 (arm64, Debian Bookworm)
- Compiled with -Wall -Wextra -Werror (no warnings)
- Hardware tested: LoRa bridge operational

Fixes:
- tap.h:46,56 error: unknown type name 'ssize_t'
- fragment.c:134 error: comparison of different signedness

+++ firmware/lora-bridge/fragment.c
@@ -131,7 +131,7 @@ int fragment_decode(const uint8_t *data, size_t len, lora_fragment_t *frag)
         return -1;
     }
 
-    if (len < FRAG_HEADER_SIZE + frag->payload_len) {
+    if (len < (size_t)(FRAG_HEADER_SIZE + frag->payload_len)) {
         fprintf(stderr, "Truncated fragment: %zu < %d\n", 
                 len, FRAG_HEADER_SIZE + frag->payload_len);
         return -1;
